backend/auth/requirements.txt
psycopg2-binary==2.9.9
pydantic[email]==2.5.0